amy_daisy cleanup before merge, also cleanup Arduino.
