+++ daisy/README.md
@@ -1,6 +1,8 @@
 # Daisy example using AMY
 
-This is a very barebones example of using AMY with an Electrosmith Daisy: https://electro-smith.com
+This is an example of using AMY with an Electrosmith Daisy: https://electro-smith.com
+
+It supports audio in, serial MIDI in, AMY sequencer.
 
 To use, set your location of AMY in the `Makefile`, then set up your Daisy dev environment, then put the board into DFU mode, and run
 
@@ -8,6 +10,7 @@ To use, set your location of AMY in the `Makefile`, then set up your Daisy dev e
 make clean && make && make program-boot; sleep 1.5; make program-dfu
 ```
 
-After it flashes, you should hear 15 Juno-6 notes out the line out port. 
+After it flashes, you should hear the AMY startup chime from the Daisy's line out.  If you have a MIDI serial interface hooked up to the Daisy (D14 USART1Rx) you should be able to play the default Juno synth on channel 1.
 
-This example could use some work to work better with the Daisy. [Please let us know if you can help!](https://github.com/shorepine/amy/issues/159)
+Known issues:
+ - There's something going wrong as a result of MIDI program changes.  These are supposed to change the AMY synth patch, and they work at first, but after around 10 program changes, notes start getting misconfigured and/or the whole thing crashes.
